BVTCK-121 Rework the test hierarchy
Note that a few tests were not extending Arquillian thus they were not
running in a container when running the tests in container mode. This
has all been fixed by this commit.

In passing, I had to run a Reorganize import on all the sources so there
might be a couple of unrelated changes but it is not that big a deal for
such a commit.